Revert "Trying to make sense of `increment_counter`"
This reverts commit a3dc13bc
