+++ core/o11y/Cargo.toml
@@ -29,7 +29,6 @@ tracing-subscriber = { version = "0.3.9", features = ["fmt", "env-filter", "regi
 bencher = "0.1.5"
 itoa = "1"
 metrics = "0.20.1"
-metrics-exporter-prometheus = "0.11.0"
 smartstring = "1.0.1"
 
 [features]
@@ -37,8 +36,4 @@ io_trace = []
 
 [[bench]]
 name = "metrics"
-harness = false
-
-[[bench]]
-name = "metrics_rs"
 harness = false
