+++ core/o11y/benches/metrics.rs
@@ -5,7 +5,6 @@ extern crate metrics;
 
 use bencher::Bencher;
 use near_o11y::metrics::{try_create_int_counter_vec, IntCounter, IntCounterVec};
-use metrics_exporter_prometheus::PrometheusBuilder;
 use once_cell::sync::Lazy;
 use once_cell::unsync::OnceCell;
 
@@ -16,30 +15,23 @@ static COUNTERS: Lazy<IntCounterVec> = Lazy::new(|| {
 const NUM_SHARDS: usize = 8;
 
 fn inc_counter_vec_with_label_values(bench: &mut Bencher) {
-    let before = COUNTERS.with_label_values(&["1"]).get();
     bench.iter(|| {
         for shard_id in 0..NUM_SHARDS {
             COUNTERS.with_label_values(&[&format!("{}", shard_id)]).inc();
         }
     });
-    let after = COUNTERS.with_label_values(&["1"]).get();
-    assert_ne!(before,after);
 }
 
 fn inc_counter_vec_with_label_values_to_string(bench: &mut Bencher) {
-    let before = COUNTERS.with_label_values(&["1"]).get();
     bench.iter(|| {
         for shard_id in 0..NUM_SHARDS {
             COUNTERS.with_label_values(&[&shard_id.to_string()]).inc();
         }
     });
-    let after = COUNTERS.with_label_values(&["1"]).get();
-    assert_ne!(before,after);
 }
 
 fn inc_counter_vec_with_label_values_smartstring(bench: &mut Bencher) {
     use std::fmt::Write;
-    let before = COUNTERS.with_label_values(&["1"]).get();
     bench.iter(|| {
         for shard_id in 0..NUM_SHARDS {
             let mut label = smartstring::alias::String::new();
@@ -47,13 +39,10 @@ fn inc_counter_vec_with_label_values_smartstring(bench: &mut Bencher) {
             COUNTERS.with_label_values(&[&label]).inc();
         }
     });
-    let after = COUNTERS.with_label_values(&["1"]).get();
-    assert_ne!(before,after);
 }
 
 fn inc_counter_vec_with_label_values_stack(bench: &mut Bencher) {
     use std::io::Write;
-    let before = COUNTERS.with_label_values(&["1"]).get();
     bench.iter(|| {
         for shard_id in 0..NUM_SHARDS {
             let mut buf = [0u8; 10];
@@ -64,12 +53,9 @@ fn inc_counter_vec_with_label_values_stack(bench: &mut Bencher) {
             COUNTERS.with_label_values(&[label]).inc();
         }
     });
-    let after = COUNTERS.with_label_values(&["1"]).get();
-    assert_ne!(before,after);
 }
 
 fn inc_counter_vec_with_label_values_stack_no_format(bench: &mut Bencher) {
-    let before = COUNTERS.with_label_values(&["1"]).get();
     bench.iter(|| {
         for shard_id in 0..NUM_SHARDS {
             let mut buf = [0u8; 16];
@@ -87,8 +73,6 @@ fn inc_counter_vec_with_label_values_stack_no_format(bench: &mut Bencher) {
             COUNTERS.with_label_values(&[label]).inc();
         }
     });
-    let after = COUNTERS.with_label_values(&["1"]).get();
-    assert_ne!(before,after);
 }
 
 fn inc_counter_vec_cached(bench: &mut Bencher) {
@@ -96,32 +80,25 @@ fn inc_counter_vec_cached(bench: &mut Bencher) {
     let counters: Vec<IntCounter> = (0..NUM_SHARDS)
         .map(|shard_id| COUNTERS.with_label_values(&[&format!("{}", shard_id)]))
         .collect();
-    let before = COUNTERS.with_label_values(&["1"]).get();
     bench.iter(|| {
         for shard_id in 0..NUM_SHARDS {
             counters[shard_id].inc();
         }
     });
-    let after = COUNTERS.with_label_values(&["1"]).get();
-    assert_ne!(before,after);
 }
 
 fn inc_counter_vec_cached_str(bench: &mut Bencher) {
     const NUM_SHARDS: usize = 8;
     let shard_ids: Vec<String> = (0..NUM_SHARDS).map(|shard_id| format!("{}", shard_id)).collect();
-    let before = COUNTERS.with_label_values(&["1"]).get();
     bench.iter(|| {
         for shard_id in 0..NUM_SHARDS {
             COUNTERS.with_label_values(&[&shard_ids[shard_id]]).inc();
         }
     });
-    let after = COUNTERS.with_label_values(&["1"]).get();
-    assert_ne!(before,after);
 }
 
 fn inc_counter_vec_itoa(bench: &mut Bencher) {
     const NUM_SHARDS: usize = 8;
-    let before = COUNTERS.with_label_values(&["1"]).get();
     bench.iter(|| {
         for shard_id in 0..NUM_SHARDS {
             let mut buffer = itoa::Buffer::new();
@@ -129,14 +106,11 @@ fn inc_counter_vec_itoa(bench: &mut Bencher) {
             COUNTERS.with_label_values(&[&printed]).inc();
         }
     });
-    let after = COUNTERS.with_label_values(&["1"]).get();
-    assert_ne!(before,after);
 }
 
 fn inc_counter_vec_cached_lazy(bench: &mut Bencher) {
     const NUM_SHARDS: usize = 8;
     let lazy_counters: Vec<OnceCell<IntCounter>> = vec![OnceCell::new(); NUM_SHARDS];
-    let before = COUNTERS.with_label_values(&["1"]).get();
     bench.iter(|| {
         for shard_id in 0..NUM_SHARDS {
             lazy_counters[shard_id]
@@ -144,8 +118,25 @@ fn inc_counter_vec_cached_lazy(bench: &mut Bencher) {
                 .inc();
         }
     });
-    let after = COUNTERS.with_label_values(&["1"]).get();
-    assert_ne!(before,after);
+}
+
+fn inc_counter_vec_metrics1(bench: &mut Bencher) {
+    const NUM_SHARDS: usize = 8;
+    bench.iter(|| {
+        for shard_id in 0..NUM_SHARDS {
+            increment_counter!("near_test_counters", "shard_id" => format!("{}",shard_id));
+        }
+    });
+}
+
+fn inc_counter_vec_metrics2(bench: &mut Bencher) {
+    const NUM_SHARDS: usize = 8;
+    bench.iter(|| {
+        for shard_id in 0..NUM_SHARDS {
+            let labels = [("shard_id", format!("{}", shard_id))];
+            increment_counter!("near_test_counters", &labels);
+        }
+    });
 }
 
 benchmark_group!(
@@ -159,5 +150,7 @@ benchmark_group!(
     inc_counter_vec_cached,
     inc_counter_vec_itoa,
     inc_counter_vec_cached_lazy,
+    inc_counter_vec_metrics1,
+    inc_counter_vec_metrics2,
 );
 benchmark_main!(benches);
