+++ integration-tests/Cargo.toml
@@ -27,8 +27,6 @@ tempfile = "3"
 tokio = { version = "1.1", features = ["net", "rt-multi-thread"] }
 tracing = "0.1.13"
 wat = "1.0"
-metrics = { package = "metrics", version = "0.20.1" }
-metrics-exporter-prometheus = "0.11.0"
 
 near-actix-test-utils = { path = "../test-utils/actix-test-utils" }
 near-chain = { path = "../chain/chain" }
